:recycle: (typescript) converted indexeddb to strict TS
