+++ upcoming-release-notes/4672.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [MatissJanis]
+---
+
+Converted indexeddb files to strict TypeScript.
